92/final order widget refactor (#183)
* Renamed OrderDetails to DetailsTable

* Renamed OrderWidgetView to OrderDetails and moved it to components/orders

* Added TODO on OrderDetails

* Added stories for DetailsTable component
