+++ src/components/orders/DetailsTable/DetailsTable.stories.tsx
@@ -1,0 +1,65 @@
+import React from 'react'
+// also exported from '@storybook/react' if you can deal with breaking changes in 6.1
+import { Story, Meta } from '@storybook/react/types-6-0'
+import BigNumber from 'bignumber.js'
+import { add, sub } from 'date-fns'
+
+import { GlobalStyles, ThemeToggler } from 'storybook/decorators'
+
+import { ONE_BIG_NUMBER } from 'const'
+
+import { DetailsTable, Props } from '.'
+
+import { RICH_ORDER } from '../../../../test/data'
+
+export default {
+  title: 'orders/DetailsTable',
+  component: DetailsTable,
+  decorators: [GlobalStyles, ThemeToggler],
+  argTypes: { order: { control: null } },
+} as Meta
+
+const Template: Story<Props> = (args) => <DetailsTable {...args} />
+
+const order = {
+  ...RICH_ORDER,
+  buyAmount: new BigNumber('1000000000000000000'), // 1WETH
+  sellAmount: new BigNumber('5000000000'), //5000 USDT
+  creationDate: sub(new Date(), { hours: 1 }),
+  expirationDate: add(new Date(), { hours: 1 }),
+}
+
+const defaultProps: Props = { order }
+
+export const DefaultFillOrKill = Template.bind({})
+DefaultFillOrKill.args = { ...defaultProps }
+
+export const FilledFillOrKill = Template.bind({})
+FilledFillOrKill.args = {
+  ...defaultProps,
+  order: {
+    ...order,
+    status: 'filled',
+    executedBuyAmount: order.buyAmount,
+    executedSellAmount: order.sellAmount,
+    filledAmount: order.sellAmount,
+    filledPercentage: ONE_BIG_NUMBER,
+  },
+}
+
+export const DefaultPartiallyFillable = Template.bind({})
+DefaultPartiallyFillable.args = { ...defaultProps, order: { ...order, partiallyFillable: true } }
+
+export const FilledPartiallyFillable = Template.bind({})
+FilledPartiallyFillable.args = {
+  ...defaultProps,
+  order: {
+    ...order,
+    partiallyFillable: true,
+    status: 'filled',
+    executedBuyAmount: order.buyAmount,
+    executedSellAmount: order.sellAmount,
+    filledAmount: order.sellAmount,
+    filledPercentage: ONE_BIG_NUMBER,
+  },
+}
